Rename vehicle ts files to adhere to other files naming conventon
