+++ lib/vehicles/american.vehicle.ts
