+++ lib/vehicles/canadian.vehicle.ts
