+++ lib/vehicles/european.vehicle.ts
